split holders in one file per class
