+++ include/Grid.hpp
@@ -1,7 +1,7 @@
 #ifndef GRID
 #define GRID
 #include "Region.hpp"
-#include "Holder.hpp"
+#include "RegionHolder.hpp"
 #include "Cell.hpp"
 #include <array>
