+++ include/RowHolder.hpp
@@ -1,0 +1,17 @@
+#ifndef ROWHOLDER_HPP
+#define ROWHOLDER_HPP
+
+#include "Cell.hpp"
+#include "TripleHolder.hpp"
+
+class RowHolder : public TripleHolder{
+public:
+        RowHolder(Cell* const g, Cell* const c, Cell* const d): TripleHolder(g,c,d){};
+
+        Cell* const G();
+        Cell* const C();
+        Cell* const D();
+};
+
+
+#endif
