+++ src/ColumnHolder.cpp
@@ -1,0 +1,6 @@
+#include "ColumnHolder.hpp"
+
+
+Cell* const ColumnHolder::T(){return getFirst();}
+Cell* const ColumnHolder::C(){return getSecond();}
+Cell* const ColumnHolder::B(){return getThird();}
