+++ src/RowHolder.cpp
@@ -1,0 +1,6 @@
+#include "RowHolder.hpp"
+
+
+Cell* const RowHolder::G(){return getFirst();}
+Cell* const RowHolder::C(){return getSecond();}
+Cell* const RowHolder::D(){return getThird();}
