+++ test/main.cpp
@@ -3,7 +3,9 @@
 
 #include "Cell.hpp"
 #include "Region.hpp"
-#include "Holder.hpp"
+#include "ColumnHolder.hpp"
+#include "RowHolder.hpp"
+#include "RegionHolder.hpp"
 #include "LastCellFinder.hpp"
 #include "Grid.hpp"
 #include "OnlyOneChoiceInColumnVisitor.hpp"
